backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
deepagents
langchain-openai
langgraph-checkpoint-sqlite>=3.0.0
aiosqlite>=0.20.0
pydantic>=2.7
python-dotenv==1.0.1
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
